drm_hwcomposer: Add simplistic SetLayerColor implementation
VTS tests for SET_LAYER_COLOR fails for now since SetLayerColor is
unsupported function. Tis commit tries to address this issue. We can't
fully implement it for now, therefore simplistic implementation that
just saves argument for future uses and satisfies the interface could
be used instead, for now.

Commit da5839cf9258 ("drm_hwcomposer: Add support for GetColorModes &
SetCursorPosition") implements GetColorModes and SetCursorPosition in
the same fashion, simply by adding fields that store this values for
future uses, therefore we assume that it is okay to use this approach
here too.

Signed-off-by: Roman Kovalivskyi <roman.kovalivskyi@globallogic.com>