drm_hwcomposer: Add GetDisplayCapabilities and getDisplayIdentificationData
Implement GetDisplayCapabilities() to get a list of supported capabilities.
The drm core doesn't support layer CTM property yet and CRTC not having the
CTM property member neither. So for this patch, we just return 0 for now.
It can fix the segment fault while booting the Android on HWC2.3, caused by
lack of this API.

Implement getDisplayIdentificationData() to get the EDID blob data
along with the size and port.
Add edid property in DrmConnector.

Signed-off-by: Lowry Li (Arm Technology China) <lowry.li@arm.com>