drm_hwcomposer: Revert "drm_hwcomposer: Extract more BO data by Imagination importer"
This reverts commit b2f2baeb9db0d0f4cbc0cac3184058d3fce85ea6.

I'm trying to utilize a policy of lighter reviews w/ quick reverts
when there are issues to keep things moving. So with that in mind
I'm reverting "Extract more BO data by Imagination importer" since
its breaking the build in AOSP.

See:
  https://android-review.googlesource.com/c/platform/external/drm_hwcomposer/+/1225155/-1..1#message-828f83ed7fa89b3bdfff7b0b5b1dd7364c822892
  https://android-review.googlesource.com/c/platform/external/drm_hwcomposer/+/1225916

This breakage is due to the importer code being tightly tied to
the gralloc private handle definitions, which is a common problem
as the gralloc and drm_hwcomposers are separately maintained.
The beagle_x15 gralloc is apparently older, and doesn't have
the newly utilized fields.

So for now, revert this to stay in sync w/ AOSP. I'm happy to
re-apply as soon as a fix is in place which adds some sort of
versioning checks so hopefully drm_hwc can work with different
versions of the gralloc code.

Once users of the old gralloc code are updated, we can remove
such version checks if needed.

Change-Id: Ie72561061ea12cb58d3ba9616a62d05f4ffe78b0
Signed-off-by: John Stultz <john.stultz@linaro.org>