drm_hwcomposer: Add rcar-du display backend
ABGR8888 pixel format and layer scaling are unsupported for rcar-du
platform.

Handle and filter out the layers by overriding the IsClientLayer method
for additional checks. That will force layers that require scaling or have
the unsupported pixel format to be merged by GPU, and allow other layers
to be merged by DRM.

Signed-off-by: Matvii Zorin <matvii.zorin@globallogic.com>